requirements.txt
asgiref==3.3.1
boto3==1.16.40
botocore==1.19.40
dj-database-url==0.5.0
Django==3.1.4
django-crispy-forms==1.10.0
django-disqus==0.5
django-filebrowser-no-grappelli==3.8.0
django-storages==1.11.1
django-tinymce4-lite==1.8.0
filebrowser-s3==1.0.0
gunicorn==20.0.4
jmespath==0.10.0
jsmin==2.2.2
mock==4.0.3
Pillow==8.0.1
python-dateutil==2.8.1
pytz==2020.4
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
psycopg2-binary
